+++ src/admin/admin.module.ts
@@ -3,9 +3,10 @@ import { AdminController } from './admin.controller';
 import { AdminService } from './admin.service';
 import { UsersEntity } from 'src/users/entities/users.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { DataAccessModule } from 'src/common/data-access.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([UsersEntity])], 
+  imports: [TypeOrmModule.forFeature([UsersEntity]), DataAccessModule], 
   controllers: [AdminController],
   providers: [AdminService],
 })
