+++ src/admin/entities/index.ts
@@ -1,0 +1,4 @@
+export * from './deleted-users.entity';
+export * from './report-comments.entity';
+export * from './report-posts.entity';
+export * from './suspended-users.entity';
